+++ lib/mvvm_service.dart
@@ -1,13 +1,10 @@
 library;
 
-export 'components/service_entry.dart';
-export 'components/service_when.dart';
+export 'components/service_mode.dart';
+export 'components/service_subscription.dart';
 export 'components/service.dart';
 
-export 'widgets/multi_service_container.dart';
-export 'widgets/multi_service_widget.dart';
-export 'widgets/service_container.dart';
-export 'widgets/service_provider.dart';
+export 'extensions/service_build_context.dart';
+export 'extensions/service_when.dart';
+
 export 'widgets/service_scope.dart';
-export 'widgets/service_widget_of.dart';
-export 'widgets/service_widget.dart';