+++ test/lib/multi_test_service.dart
@@ -1,9 +1,0 @@
-import 'package:mvvm_service/mvvm_service.dart';
-
-import 'test_service.dart';
-
-/// A test implementation of [Service] for multi-service scenarios.
-class MultiTestService1 extends TestService {}
-
-/// A test implementation of [Service] for multi-service scenarios.
-class MultiTestService2 extends TestService {}