+++ test/lib/test_service.dart
@@ -7,6 +7,7 @@ class TestService extends Service<String> {
   TestService({this.isThrowError = false});
 
   final bool isThrowError;
+  bool isDisposed = false;
 
   static const Duration duration = Duration(microseconds: 1);
   static const String sampleData = "Hello, World!";
@@ -19,4 +20,14 @@ class TestService extends Service<String> {
     await Future.delayed(Duration(microseconds: 1));
     return sampleData;
   }
+
+  @override
+  void dispose() {
+    isDisposed = true;
+    super.dispose();
+  }
 }
+
+class TestService1 extends TestService {}
+
+class TestService2 extends TestService {}